requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.3
django-crispy-forms==2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
